HUE-457 Cannot delete directories with spaces in the name
The filebrowser app could not delete files or directories with spaces in the name because
the back-end would have already escaped the URL, and so the second escaping on the
client-side when attempting to send an AJAX request for the deletion would result in
a broken link. The fix was to patch up reverse_with_get in django_util so that it
offers an option to not url encode the GET parameters on the server side.